refactor: move all notifications functions under the same module
